Ability to display well formatter longer description in talks page
By default content is used as a subtitle - displayed right under Title
without extra formatting (to keep backward compatibility). However,
it doesn't look good to longer (multiline) content (margins, etc.).
For that cases there is an option to define the Subtitle parameter in _index.md
to render is as a subtitle and then .Content is rendered within the main tag
with better formatting.